Replace long int with intmax_t
`intmax_t` is closer to the XLEN sized integer that the SBI spec means
when using `long`, but I would still prefer to avoid dependency on the C
specification as it can only bring problems.

The C specification defines `long` as implementation specific with
minimal width of 32 bits.  It just so happens that Unixes use `long`
that matches XLEN for 32 and 64 bit systems.
If RISC-V happened to be popular enough to attract Microsoft, MSVC would
try to have 32 bit `long` on XLEN=64.

The situation will get even more complicated with RV128I and XLEN=128,
because I think even Unixes would keep `long` to 64 bits and either bump
`long long` or keep it available only as int128_t.

[The full version of this patch would replace all `long`s in the spec,
 but I just wanted to get the point across.]