Clarify reserved bits
Not defining reserved bits on input can make future extensions harder as
we cannot just define new functionality for the bit, but have to add a
toggle feature that first enables the bit.
This is because the supervisor implementation could have provided any
value for the reserved bits and we cannot implicitly expect zero.

The downside is that this slows down the interface by a few cycles,
especially since most functions don't use all extra registers in `a0` to
`a5` range, but the supervisor would have to check them for zero.

We already decided that the overhead of clobbering the unspecified
registers is acceptable and the extra check for zero has less overhead
the the stack manipulation that resulted from previous decision.